+++ src/android/Receiver.java
@@ -35,8 +35,6 @@ import android.app.PendingIntent;
 import android.content.BroadcastReceiver;
 import android.content.Context;
 import android.content.Intent;
-import android.graphics.Bitmap;
-import android.graphics.BitmapFactory;
 import android.net.Uri;
 import android.os.Build;
 import android.os.Bundle;
@@ -118,8 +116,7 @@ public class Receiver extends BroadcastReceiver {
      */
     @SuppressLint("NewApi")
     private Builder buildNotification () {
-        Bitmap icon = BitmapFactory.decodeResource(context.getResources(), options.getIcon());
-        Uri sound   = options.getSound();
+        Uri sound = options.getSound();
 
         Builder notification = new NotificationCompat.Builder(context)
             .setContentTitle(options.getTitle())
@@ -127,7 +124,7 @@ public class Receiver extends BroadcastReceiver {
             .setNumber(options.getBadge())
             .setTicker(options.getMessage())
             .setSmallIcon(options.getSmallIcon())
-            .setLargeIcon(icon)
+            .setLargeIcon(options.getIcon())
             .setAutoCancel(options.getAutoCancel())
             .setOngoing(options.getOngoing());